[minor update] fix(security): rate-limit login and keep fractional local-position nudges
The login endpoint locks a client after repeated failures in a window and answers a locked client with 429 and Retry-After, so the internet-reachable station is not an open brute-force target. The lockout logic is a pure, unit-tested helper. The set_position_local route parses its meters as floats instead of integers, so a fractional-meter nudge keeps its decimals through to the vehicle.

+++ compose/svelte-kit/src/routes/api/mavlink/[type]/+server.ts
@@ -116,9 +116,9 @@ export const POST: RequestHandler = async (event): Promise<Response> => {
             }
         }
         case 'set_position_local': {
-            const x: number = parseInt(event.request.headers.get('x')!);
-            const y: number = parseInt(event.request.headers.get('y')!);
-            const z: number = parseInt(event.request.headers.get('z')!);
+            const x: number = parseFloat(event.request.headers.get('x')!);
+            const y: number = parseFloat(event.request.headers.get('y')!);
+            const z: number = parseFloat(event.request.headers.get('z')!);
             if (isNaN(x) || isNaN(y) || isNaN(z)) {
                 return new Response('Invalid coordinates', { status: 400 });
             }
